Add files via upload
Attempted to include dendograms with ggplot heat maps but the ordering expressed by the dendograms does not reflect in the heat map, when compared with the heatmap generated using the 'heatmap' function. In this plot however, I am unable to move the legend outside the plot region, causing it to hinder the view of the heat-map. It is necessary to include a legend as it allows one to understand the meaning of the different colors in the heat-map.